Refactor auth reducer
