+++ src/redux/actions/reservations.js
@@ -1,12 +1,11 @@
 import { GET_MY_RESERVATIONS } from '.';
-import { getToken } from './auth';
+import { headers, getToken } from '../../utils';
 
 const getReservations = () => async (dispatch) => {
   const response = await fetch('http://localhost:3001/v1/reservations', {
     method: 'GET',
     headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
+      ...headers,
       Authorization: getToken(),
     },
   });
